apply clean architecture
