+++ app/src/main/java/com/jys/catsapp/data/database/PhotoRemoteKeyDao.kt
@@ -1,17 +1,18 @@
-package com.jys.catsapp.data.localDB
+package com.jys.catsapp.data.database
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import com.jys.catsapp.data.database.entity.PhotoRemoteKeyEntity
 
 @Dao
 interface PhotoRemoteKeyDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertOrReplace(remoteKey: PhotoRemoteKey)
+    suspend fun insertOrReplace(remoteKey: PhotoRemoteKeyEntity)
 
     @Query("SELECT * FROM photo_remote_keys WHERE photoId = :photoId")
-    suspend fun remoteKeyByPhoto(photoId: String): PhotoRemoteKey?
+    suspend fun remoteKeyByPhoto(photoId: String): PhotoRemoteKeyEntity?
 
     @Query("DELETE FROM photo_remote_keys")
     suspend fun clearRemoteKeys()
