+++ app/src/main/java/com/jys/catsapp/data/database/entity/PhotoRemoteKeyEntity.kt
@@ -1,10 +1,10 @@
-package com.jys.catsapp.data.localDB
+package com.jys.catsapp.data.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "photo_remote_keys")
-data class PhotoRemoteKey(
+data class PhotoRemoteKeyEntity(
     @PrimaryKey val photoId: String,
     val prevPageKey: Int?,
     val nextPageKey: Int?
