+++ app/src/main/java/com/jys/catsapp/di/databaseModule.kt
@@ -1,7 +1,7 @@
 package com.jys.catsapp.di
 
 import androidx.room.Room
-import com.jys.catsapp.data.localDB.CatsDatabase
+import com.jys.catsapp.data.database.CatsDatabase
 import org.koin.dsl.module
 
 val databaseModule = module {
