+++ app/src/main/java/com/jys/catsapp/domain/model/PhotoDomain.kt
@@ -1,7 +1,6 @@
 package com.jys.catsapp.domain.model
 
-import com.jys.catsapp.data.localDB.PhotoEntity
-import com.jys.catsapp.data.localDB.SrcEntity
+import com.jys.catsapp.data.database.entity.PhotoEntity
 
 data class PhotoDomain(
     val id: Int?,
@@ -16,18 +15,6 @@ data class PhotoDomain(
     val alt: String?,
     val liked: Boolean?
 )
-
-data class SrcDomain(
-    val original: String?,
-    val large2x: String?,
-    val large: String?,
-    val medium: String?,
-    val small: String?,
-    val portrait: String?,
-    val landscape: String?,
-    val tiny: String?
-)
-
 
 fun PhotoEntity.toDomain(): PhotoDomain {
     return PhotoDomain(
@@ -54,17 +41,4 @@ fun PhotoEntity.toDomain(): PhotoDomain {
     )
 }
 
-
-fun SrcEntity.toDomain(): SrcDomain {
-    return SrcDomain(
-        original = this.original,
-        large2x = this.large2x,
-        large = this.large,
-        medium = this.medium,
-        small = this.small,
-        portrait = this.portrait,
-        landscape = this.landscape,
-        tiny = this.tiny
-    )
-}
