+++ gradle/libs.versions.toml
@@ -11,7 +11,6 @@ composeBom = "2024.06.00"
 ksp = '1.9.0-1.0.12'
 
 exoplayer = "2.18.0"
-media3 = "1.0.0"
 
 navigationCompose = "2.7.7"
 coilCompose = "2.6.0"
